Use javascript-action instead of platform dependent Shell commands
Instead of a simple "test -x eldev" test, eldev is actually
bootstrapped again to verify that the net-installation was indeed
successful.

The previous shell tests also used the hard-coded path
https://raw.github.com/doublep/eldev/master/bin/eldev to download the
eldev scripts. So CI newer tested changes introduced in PRs.

The utility function getRawUrl generates the correct raw download
URL depending on the context (pushing to master or PR) and also adds a
platform dependent suffix (.bat on Windows) to the script name to make
the script download/testing platform-agnostic.